requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
